# openapi.json
{
  "components": {
    "schemas": {
      "CreateQueryDto": {
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 200,
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "organization": {
            "maxLength": 200,
            "type": "string"
          },
          "requirements": {
            "maxLength": 5000,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "requirements"
        ],
        "type": "object"
      },
      "CreateQueryResponseDto_Output": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt"
        ],
        "type": "object"
      },
      "ErrorResponseDto": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "example": "VALIDATION_ERROR",
                "type": "string"
              },
              "details": {},
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "HealthResponseDto_Output": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "enum": [
              "ok"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "QueryDto_Output": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "organization": {
            "nullable": true,
            "type": "string"
          },
          "requirements": {
            "type": "string"
          },
          "status": {
            "enum": [
              "NEW",
              "IN_PROGRESS",
              "RESOLVED"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "organization",
          "email",
          "requirements",
          "status",
          "createdAt"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "adminApiKey": {
        "bearerFormat": "API key",
        "description": "Temporary admin API key (until Stage 2 auth)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Kritex ecommerce API. Errors always use `{ \"error\": { \"code\", \"message\", \"details\"? } }`.",
    "title": "Kritex API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto_Output"
                }
              }
            },
            "description": "API and database are up"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Database unavailable"
          }
        },
        "summary": "Liveness + database check",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/queries": {
      "get": {
        "operationId": "listQueries",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QueryDto_Output"
                  },
                  "type": "array"
                }
              }
            },
            "description": "All queries"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "adminApiKey": []
          }
        ],
        "summary": "List submitted inquiries, newest first (admin)",
        "tags": [
          "queries"
        ]
      },
      "post": {
        "operationId": "createQuery",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQueryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateQueryResponseDto_Output"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "VALIDATION_ERROR"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "TOO_MANY_REQUESTS"
          }
        },
        "summary": "Submit a contact / tender inquiry (public)",
        "tags": [
          "queries"
        ]
      }
    }
  },
  "servers": [],
  "tags": []
}